backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
openai-whisper
torch
torchaudio
numpy
requests 